Move llvm.h into src/backends/llvm/ subdir
