+++ src/backends/llvm/llvm.c
@@ -5,7 +5,7 @@
 #pragma GCC diagnostic pop
 #include <llvm-c/ExecutionEngine.h>
 #include "../../internal.h"
-#include "../../llvm.h"
+#include "llvm.h"
 
 typedef struct HLLVMParser_ {
   LLVMModuleRef mod;
